v5/requirements.txt
# Dependencies
flask
psycopg2-binary
spacy
requests
python-dotenv